+++ CHANGES.md
@@ -8,6 +8,7 @@ git version
       files (#1219 by @ddickstein, fixes #894)
     - fix handling of ppx's under Windows (#1413)
     - handle `=` syntax in compiler flags (#1409)
+    - fix superfluous break in error reporting (#1432)
   + editor modes
     - update quick setup instructions for emacs (#1380, @ScriptDevil)
     - fix an issue in Neovim where the current line jumps to the top of the